db/requirements.txt
mysql-connector-python==8.3.0
python-dotenv==1.0.1
